refactor: onboarding project tree
