+++ ts/features/onboarding/hooks/useOnboardingAbortAlert.ts
@@ -1,8 +1,8 @@
 import { Alert } from "react-native";
 import { useCallback } from "react";
-import { abortOnboarding } from "../../store/actions/onboarding";
-import I18n from "../../i18n";
-import { useIODispatch } from "../../store/hooks";
+import { abortOnboarding } from "../store/actions";
+import I18n from "../../../i18n";
+import { useIODispatch } from "../../../store/hooks";
 
 type OnboardingAbortAlertUtils = {
   showAlert: (callback?: () => void) => void;
