+++ ts/features/onboarding/saga/completeOnboardingSaga.ts
@@ -1,8 +1,8 @@
 import { call, take } from "typed-redux-saga/macro";
 import { CommonActions, StackActions } from "@react-navigation/native";
-import NavigationService from "../../navigation/NavigationService";
-import ROUTES from "../../navigation/routes";
-import { completeOnboarding } from "../../store/actions/onboarding";
+import NavigationService from "../../../navigation/NavigationService";
+import ROUTES from "../../../navigation/routes";
+import { completeOnboarding } from "../store/actions";
 
 export function* completeOnboardingSaga() {
   yield* call(
