+++ ts/features/onboarding/screens/OnboardingCompletedScreen.tsx
@@ -1,15 +1,15 @@
 import { useRoute } from "@react-navigation/native";
 import * as O from "fp-ts/lib/Option";
-import { isFastLoginEnabledSelector } from "../../features/authentication/fastLogin/store/selectors";
-import I18n from "../../i18n";
-import { completeOnboarding } from "../../store/actions/onboarding";
-import { useIODispatch, useIOSelector } from "../../store/hooks";
-import { idpSelector } from "../../features/authentication/common/store/selectors";
-import { getFlowType } from "../../utils/analytics";
-import { useOnFirstRender } from "../../utils/hooks/useOnFirstRender";
-import { trackThankYouPageScreen } from "../../features/settings/common/analytics";
-import { OperationResultScreenContent } from "../../components/screens/OperationResultScreenContent";
-import { trackLoginEnded } from "../../features/authentication/common/analytics";
+import { isFastLoginEnabledSelector } from "../../authentication/fastLogin/store/selectors";
+import I18n from "../../../i18n";
+import { completeOnboarding } from "../store/actions";
+import { useIODispatch, useIOSelector } from "../../../store/hooks";
+import { idpSelector } from "../../authentication/common/store/selectors";
+import { getFlowType } from "../../../utils/analytics";
+import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
+import { trackThankYouPageScreen } from "../../settings/common/analytics";
+import { OperationResultScreenContent } from "../../../components/screens/OperationResultScreenContent";
+import { trackLoginEnded } from "../../authentication/common/analytics";
 
 const OnboardingCompletedScreen = () => {
   const dispatch = useIODispatch();
