+++ ts/features/onboarding/screens/__tests__/OnboardingTosScreen.test.tsx
@@ -1,30 +1,30 @@
 import * as pot from "@pagopa/ts-commons/lib/pot";
 import * as O from "fp-ts/lib/Option";
 import configureMockStore from "redux-mock-store";
-import I18n from "../../../i18n";
+import I18n from "../../../../i18n";
 // import WebView from "react-native-webview";
 // import {
 //   WebViewErrorEvent,
 //   WebViewNavigationEvent
 // } from "react-native-webview/lib/WebViewTypes";
-import { appReducer } from "../../../store/reducers";
-import { applicationChangeState } from "../../../store/actions/application";
-import { GlobalState } from "../../../store/reducers/types";
-import { ToolEnum } from "../../../../definitions/content/AssistanceToolConfig";
-import { InitializedProfile } from "../../../../definitions/backend/InitializedProfile";
-import ROUTES from "../../../navigation/routes";
-import { renderScreenWithNavigationStoreContext } from "../../../utils/testWrapper";
+import { appReducer } from "../../../../store/reducers";
+import { applicationChangeState } from "../../../../store/actions/application";
+import { GlobalState } from "../../../../store/reducers/types";
+import { ToolEnum } from "../../../../../definitions/content/AssistanceToolConfig";
+import { InitializedProfile } from "../../../../../definitions/backend/InitializedProfile";
+import ROUTES from "../../../../navigation/routes";
+import { renderScreenWithNavigationStoreContext } from "../../../../utils/testWrapper";
 // import * as ToastUtils from "../../../utils/showToast";
 import OnboardingTosScreen from "../OnboardingTosScreen";
-import { ServicesPreferencesModeEnum } from "../../../../definitions/backend/ServicesPreferencesMode";
-import { mockAccessibilityInfo } from "../../../utils/testAccessibility";
+import { ServicesPreferencesModeEnum } from "../../../../../definitions/backend/ServicesPreferencesMode";
+import { mockAccessibilityInfo } from "../../../../utils/testAccessibility";
 
 const CurrentTestToSVersion = 2.0;
 
 // Restore defineProperty
 beforeAll(() => {
   jest.resetAllMocks();
-  jest.mock("./../../../config");
+  jest.mock("../../../../config");
   mockAccessibilityInfo(false);
 });
