+++ ts/features/onboarding/screens/biometric&securityChecks/MissingDeviceBiometricScreen.tsx
@@ -1,17 +1,17 @@
 import { ListItemInfo } from "@pagopa/io-app-design-system";
 import { useMemo } from "react";
-import { ContextualHelpPropsMarkdown } from "../../../components/screens/BaseScreenComponent";
-import { useHeaderSecondLevel } from "../../../hooks/useHeaderSecondLevel";
-import I18n from "../../../i18n";
-import { preferenceFingerprintIsEnabledSaveSuccess } from "../../../store/actions/persistedPreferences";
-import { useIODispatch, useIOSelector } from "../../../store/hooks";
-import { isProfileFirstOnBoardingSelector } from "../../../features/settings/common/store/selectors";
-import { getFlowType } from "../../../utils/analytics";
-import { FAQsCategoriesType } from "../../../utils/faq";
-import { useOnboardingAbortAlert } from "../../../utils/hooks/useOnboardingAbortAlert";
-import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
-import { IOScrollViewWithListItems } from "../../../components/ui/IOScrollViewWithListItems";
-import { IOScrollViewActions } from "../../../components/ui/IOScrollView";
+import { ContextualHelpPropsMarkdown } from "../../../../components/screens/BaseScreenComponent";
+import { useHeaderSecondLevel } from "../../../../hooks/useHeaderSecondLevel";
+import I18n from "../../../../i18n";
+import { preferenceFingerprintIsEnabledSaveSuccess } from "../../../../store/actions/persistedPreferences";
+import { useIODispatch, useIOSelector } from "../../../../store/hooks";
+import { isProfileFirstOnBoardingSelector } from "../../../settings/common/store/selectors";
+import { getFlowType } from "../../../../utils/analytics";
+import { FAQsCategoriesType } from "../../../../utils/faq";
+import { useOnboardingAbortAlert } from "../../hooks/useOnboardingAbortAlert";
+import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender";
+import { IOScrollViewWithListItems } from "../../../../components/ui/IOScrollViewWithListItems";
+import { IOScrollViewActions } from "../../../../components/ui/IOScrollView";
 import { trackBiometricConfigurationEducationalScreen } from "./analytics";
 
 const FAQ_CATEGORIES: ReadonlyArray<FAQsCategoriesType> = [
