+++ ts/features/onboarding/screens/biometric&securityChecks/MissingDevicePinScreen.tsx
@@ -1,18 +1,18 @@
 import { ListItemInfo } from "@pagopa/io-app-design-system";
 import { useMemo } from "react";
 import { Platform } from "react-native";
-import { ContextualHelpPropsMarkdown } from "../../../components/screens/BaseScreenComponent";
-import I18n from "../../../i18n";
-import { preferenceFingerprintIsEnabledSaveSuccess } from "../../../store/actions/persistedPreferences";
-import { useIODispatch, useIOSelector } from "../../../store/hooks";
-import { isProfileFirstOnBoardingSelector } from "../../../features/settings/common/store/selectors";
-import { getFlowType } from "../../../utils/analytics";
-import { useOnFirstRender } from "../../../utils/hooks/useOnFirstRender";
-import { useOnboardingAbortAlert } from "../../../utils/hooks/useOnboardingAbortAlert";
-import { useHeaderSecondLevel } from "../../../hooks/useHeaderSecondLevel";
-import { FAQsCategoriesType } from "../../../utils/faq";
-import { IOScrollViewWithListItems } from "../../../components/ui/IOScrollViewWithListItems";
-import { IOScrollViewActions } from "../../../components/ui/IOScrollView";
+import { ContextualHelpPropsMarkdown } from "../../../../components/screens/BaseScreenComponent";
+import I18n from "../../../../i18n";
+import { preferenceFingerprintIsEnabledSaveSuccess } from "../../../../store/actions/persistedPreferences";
+import { useIODispatch, useIOSelector } from "../../../../store/hooks";
+import { isProfileFirstOnBoardingSelector } from "../../../settings/common/store/selectors";
+import { getFlowType } from "../../../../utils/analytics";
+import { useOnFirstRender } from "../../../../utils/hooks/useOnFirstRender";
+import { useOnboardingAbortAlert } from "../../hooks/useOnboardingAbortAlert";
+import { useHeaderSecondLevel } from "../../../../hooks/useHeaderSecondLevel";
+import { FAQsCategoriesType } from "../../../../utils/faq";
+import { IOScrollViewWithListItems } from "../../../../components/ui/IOScrollViewWithListItems";
+import { IOScrollViewActions } from "../../../../components/ui/IOScrollView";
 import { trackPinEducationalScreen } from "./analytics";
 
 const FAQ_CATEGORIES: ReadonlyArray<FAQsCategoriesType> = [
