+++ ts/features/onboarding/store/selectors/index.ts
@@ -1,0 +1,5 @@
+import { GlobalState } from "../../../../store/reducers/types";
+
+export const isFingerprintAcknowledgedSelector = (
+  state: GlobalState
+): boolean => state.onboarding.isFingerprintAcknowledged;
