+++ ts/features/settings/security/shared/components/PinCreation.tsx
@@ -30,7 +30,7 @@ import { setAccessibilityFocus } from "../../../../../utils/accessibility";
 import { getFlowType } from "../../../../../utils/analytics";
 import { PIN_LENGTH_SIX } from "../../../../../utils/constants";
 import { isDevEnv } from "../../../../../utils/environment";
-import { useOnboardingAbortAlert } from "../../../../../utils/hooks/useOnboardingAbortAlert";
+import { useOnboardingAbortAlert } from "../../../../onboarding/hooks/useOnboardingAbortAlert";
 import { useOnFirstRender } from "../../../../../utils/hooks/useOnFirstRender";
 import { usePreventScreenCapture } from "../../../../../utils/hooks/usePreventScreenCapture";
 import { Carousel } from "../../../../../components/Carousel";
