+++ ts/sagas/services/servicesOptinSaga.ts
@@ -1,7 +1,7 @@
 import { call, select, take } from "typed-redux-saga/macro";
 import { StackActions } from "@react-navigation/native";
 import { navigateToServicesPreferenceModeSelectionScreen } from "../../store/actions/navigation";
-import { servicesOptinCompleted } from "../../store/actions/onboarding";
+import { servicesOptinCompleted } from "../../features/onboarding/store/actions";
 import { profileServicePreferencesModeSelector } from "../../features/settings/common/store/selectors";
 import { isServicesPreferenceModeSet } from "../../features/settings/common/store/utils/guards";
 import NavigationService from "../../navigation/NavigationService";
