+++ ts/store/reducers/types.ts
@@ -12,6 +12,7 @@ import { ProfileState } from "../../features/settings/common/store/reducers";
 import { UserDataProcessingState } from "../../features/settings/common/store/reducers/userDataProcessing";
 import { EmailValidationState } from "../../features/mailCheck/store/reducers/emailValidation";
 import { PersistedIdentificationState } from "../../features/identification/store/reducers";
+import { OnboardingState } from "../../features/onboarding/store/reducers";
 import { AppState } from "./appState";
 import { AssistanceToolsState } from "./assistanceTools";
 import { ContentState } from "./content";
@@ -20,7 +21,6 @@ import { PersistedDebugState } from "./debug";
 import { PersistedEntitiesState } from "./entities";
 import { InstallationState } from "./installation";
 import { NavigationState } from "./navigation";
-import { OnboardingState } from "./onboarding";
 import { PersistedPreferencesState } from "./persistedPreferences";
 import { PreferencesState } from "./preferences";
 import { SearchState } from "./search";
